[FLINK-19822][table-planner] Do some code cleanup for BatchExecCalcBase and BatchExecCorrelateBase
